tracer/requirements.txt
asgiref==3.2.10
certifi==2020.4.5.2
chardet==3.0.4
cos-python-sdk-v5==1.8.0
dicttoxml==1.7.4
Django==1.11.28
django-redis==4.11.0
idna==2.9
Pillow==7.0.0
pycryptodome==3.9.8
pytz==2020.1
qcloud-python-sts==3.0.3
qcloudsms-py==0.1.4
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wincertstore==0.2
